Fixed episodes with decimals not working - Reset the whole Database

+++ app/src/main/java/com/elfennani/aniwatch/data/local/entities/CachedEpisodeDto.kt
@@ -11,7 +11,7 @@ data class CachedEpisodeDto(
     @PrimaryKey val id: String,
     val allanimeId: String,
     val animeId: Int,
-    val episode: Int,
+    val episode: Double,
     val name: String,
     val dubbed: Boolean,
     val thumbnail: String?,
