+++ app/src/main/java/com/elfennani/aniwatch/data/remote/models/NetworkEpisode.kt
@@ -9,7 +9,7 @@ data class NetworkEpisode(
     val id: String,
     val allanimeId: String,
     val animeId: Int,
-    val episode: Int,
+    val episode: Double,
     val name: String,
     val dubbed: Boolean,
     val thumbnail: String?,
