+++ app/src/main/java/com/elfennani/aniwatch/ui/screens/show/ShowViewModel.kt
@@ -3,40 +3,29 @@ package com.elfennani.aniwatch.ui.screens.show
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import androidx.core.os.bundleOf
-import androidx.datastore.migrations.SharedPreferencesView
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
-import com.elfennani.aniwatch.R
-import com.elfennani.aniwatch.data.local.dao.DownloadDao
-import com.elfennani.aniwatch.data.local.entities.LocalDownloadState
-import com.elfennani.aniwatch.data.local.entities.LocalDownloadedEpisode
 import com.elfennani.aniwatch.data.repository.DownloadRepository
 import com.elfennani.aniwatch.data.repository.ShowRepository
 import com.elfennani.aniwatch.dataStore
-import com.elfennani.aniwatch.models.DownloadStatus
 import com.elfennani.aniwatch.models.EpisodeAudio
-import com.elfennani.aniwatch.models.EpisodeState
 import com.elfennani.aniwatch.models.Resource
 import com.elfennani.aniwatch.services.DownloadService
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.combineTransform
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import java.io.File
-import java.util.Date
 import javax.inject.Inject
 
 @HiltViewModel
@@ -85,7 +74,7 @@ class ShowViewModel @Inject constructor(
         }
     }
 
-    fun deleteEpisode(episode: Int) {
+    fun deleteEpisode(episode: Double) {
         viewModelScope.launch { downloadRepository.deleteDownload(showId, episode) }
             .invokeOnCompletion {
                 val directory = File(context.filesDir, "shows/$showId")
@@ -95,7 +84,7 @@ class ShowViewModel @Inject constructor(
             }
     }
 
-    fun downloadEpisode(episode: Int, audio: EpisodeAudio) {
+    fun downloadEpisode(episode: Double, audio: EpisodeAudio) {
         viewModelScope.launch {
             downloadRepository.addDownload(showId, episode, audio)
         }
